+++ app/store/stores/search-word.ts
