+++ app/store/stores/taken-lecture-atom-hydrator.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { takenLectureAtom } from '@/app/store/custom-taken-lecture';
+import { takenLectureAtom } from '@/app/store/stores/custom-taken-lecture';
 import { TakenLectrueInfo } from '@/app/type/lecture';
 import { useHydrateAtoms } from 'jotai/utils';
