+++ app/ui/lecture/lecture-search/lecture-search-result.tsx
@@ -1,6 +1,6 @@
 import { SearchedLectureInfo } from '@/app/type/lecture';
 import { useAtomValue } from 'jotai';
-import { searchWordAtom } from '@/app/store/search-word';
+import { searchWordAtom } from '@/app/store/stores/search-word';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { fetchSearchLectures } from '@/app/business/lecture/search-lecture.query';
 import List from '@/app/ui/view/molecule/list';
