+++ app/ui/lecture/taken-lecture/index.tsx
@@ -1,7 +1,7 @@
 import { fetchTakenLectures } from '@/app/business/lecture/taken-lecture.query';
 import TakenLectureList from './taken-lecture-list';
 import TakenLectureLabel from './taken-lecture-label';
-import TakenLectureAtomHydrator from '@/app/store/taken-lecture-atom-hydrator';
+import TakenLectureAtomHydrator from '@/app/store/stores/taken-lecture-atom-hydrator';
 import UpdateTakenLecture from './update-taken-lecture';
 
 export default async function TakenLecture() {
