+++ app/ui/lecture/taken-lecture/taken-lecture.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { expect, userEvent } from '@storybook/test';
 import { mockDatabase, resetMockDB } from '@/app/mocks/db.mock';
 import TakenLectureList from './taken-lecture-list';
-import TakenLectureAtomHydrator from '@/app/store/taken-lecture-atom-hydrator';
+import TakenLectureAtomHydrator from '@/app/store/stores/taken-lecture-atom-hydrator';
 import { screen } from '@storybook/testing-library';
 import { delay } from 'msw';
