+++ app/ui/result/result-category-card/result-category-card.tsx
@@ -14,7 +14,7 @@ import {
 import { DIALOG_KEY } from '@/app/utils/key/dialog-key.util';
 import Link from 'next/link';
 import { useSetAtom } from 'jotai';
-import { isDialogOpenAtom } from '@/app/store/dialog';
+import { isDialogOpenAtom } from '@/app/store/stores/dialog';
 import Button from '../../view/atom/button/button';
 import { getPercentage } from '@/app/utils/chart.util';
 import PieChart from '../../view/molecule/pie-chart/pie-chart';
